Remove Influxdb usage and move to using Data type
The current dashboard setup no longer uses influxdb in any capacity, so any code relating to it is dead code. The move to PostgreSQL was motivated by performance issues with influxdb (at least in the specific
use case of ours with Grafana).

The new Data type is introduced to better facilitate changes to the dashboard. Rather than *update* the schema for the dashboard_data table with a new column and do a whole bunch of database UPDATEs it seems that it would be more performant to simply create a new table in PostgreSQL and do a separate insertion, while updating all pre-existing json files.

JSON files are now versioned to reflect changes to the database setup.